#168 use swift additions as precombiled xcframework
