+++ mvvm-flow/apple/src/commonMain/kotlin/Greeting.kt
@@ -1,0 +1,8 @@
+/*
+ * Copyright 2022 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package dev.icerock.moko.mvvm.flow.apple
+
+// just for compile this module
+fun helloWorld() = println("hello world")
