+++ mvvm-flow/apple/xcode/.gitignore
@@ -1,0 +1,1 @@
+xcframework/
