+++ mvvm-flow/apple/xcode/Shared/ContentView.swift
@@ -1,0 +1,21 @@
+//
+//  ContentView.swift
+//  Shared
+//
+//  Created by Aleksey Mikhailov on 29.04.2022.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    var body: some View {
+        Text("Hello, world!")
+            .padding()
+    }
+}
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        ContentView()
+    }
+}
