+++ mvvm-flow/apple/xcode/Shared/mokoMvvmFlowApp.swift
@@ -1,0 +1,17 @@
+//
+//  mokoMvvmFlowApp.swift
+//  Shared
+//
+//  Created by Aleksey Mikhailov on 29.04.2022.
+//
+
+import SwiftUI
+
+@main
+struct mokoMvvmFlowApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
