+++ mvvm-flow/apple/xcode/mokoMvvmFlowSwiftUI/ViewModelObservable.swift
@@ -1,0 +1,13 @@
+//
+//  CFlowExt.swift
+//  mokoMvvmFlowSwiftUI (iOS)
+//
+//  Created by Aleksey Mikhailov on 29.04.2022.
+//
+
+import MultiPlatformLibrary
+import SwiftUI
+
+extension ViewModel: ObservableObject {
+    
+}
