+++ mvvm-flow/apple/xcode/mokoMvvmFlowSwiftUI/mokoMvvmFlowSwiftUI.h
@@ -1,0 +1,18 @@
+//
+//  mokoMvvmFlowSwiftUI.h
+//  mokoMvvmFlowSwiftUI
+//
+//  Created by Aleksey Mikhailov on 29.04.2022.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for mokoMvvmFlowSwiftUI.
+FOUNDATION_EXPORT double mokoMvvmFlowSwiftUIVersionNumber;
+
+//! Project version string for mokoMvvmFlowSwiftUI.
+FOUNDATION_EXPORT const unsigned char mokoMvvmFlowSwiftUIVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <mokoMvvmFlowSwiftUI/PublicHeader.h>
+
+
