+++ settings.gradle.kts
@@ -17,6 +17,7 @@ includeBuild("mvvm-build-logic")
 include(":mvvm-internal")
 include(":mvvm-core")
 include(":mvvm-flow")
+include(":mvvm-flow:apple")
 include(":mvvm-flow-compose")
 include(":mvvm-livedata")
 include(":mvvm-livedata-material")
